requirements.txt
certifi==2019.3.9
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap3==11.0.0
django-cors-headers==3.0.0
django-filter==2.1.0
django-heroku==0.3.1
django-mailgun==0.9.1
django-nose==1.4.6
django-recaptcha==2.0.2
django-tinymce==2.8.0
djangorestframework==3.9.4
djangorestframework-datatables==0.5.0
idna==2.8
Markdown==3.1.1
nose==1.3.7
psycopg2==2.8.3
pytz==2019.1
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
tmdbsimple==2.2.0
urllib3==1.25.3
whitenoise==4.1.2
